refactored expand folder to be used everywhere without circular imports
