+++ cmd/random.go
@@ -13,6 +13,7 @@ import (
 	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator"
 	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/models"
 	"github.com/krkn-chaos/krknctl/pkg/scenario_orchestrator/utils"
+	commonutils "github.com/krkn-chaos/krknctl/pkg/utils"
 	"github.com/spf13/cobra"
 	"log"
 	"os"
@@ -62,7 +63,7 @@ func NewRandomRunCommand(factory *providerfactory.ProviderFactory, scenarioOrche
 				return err
 			}
 			if kubeconfig != "" {
-				expandedConfig, err := utils.ExpandFolder(kubeconfig, nil)
+				expandedConfig, err := commonutils.ExpandFolder(kubeconfig, nil)
 				if err != nil {
 					return err
 				}
@@ -76,7 +77,7 @@ func NewRandomRunCommand(factory *providerfactory.ProviderFactory, scenarioOrche
 				return err
 			}
 			if alertsProfile != "" {
-				expandedProfile, err := utils.ExpandFolder(alertsProfile, nil)
+				expandedProfile, err := commonutils.ExpandFolder(alertsProfile, nil)
 				if err != nil {
 					return err
 				}
@@ -90,7 +91,7 @@ func NewRandomRunCommand(factory *providerfactory.ProviderFactory, scenarioOrche
 				return err
 			}
 			if metricsProfile != "" {
-				expandedProfile, err := utils.ExpandFolder(metricsProfile, nil)
+				expandedProfile, err := commonutils.ExpandFolder(metricsProfile, nil)
 				if err != nil {
 					return err
 				}
@@ -308,7 +309,7 @@ func NewRandomScaffoldCommand(factory *providerfactory.ProviderFactory, config c
 			}
 
 			if seedFile != "" {
-				seedFilePath, err := utils.ExpandFolder(seedFile, nil)
+				seedFilePath, err := commonutils.ExpandFolder(seedFile, nil)
 				if err != nil {
 					return err
 				}
