+++ pkg/scenario_orchestrator/utils/utils_test.go
@@ -8,7 +8,6 @@ import (
 	"io/fs"
 	"k8s.io/client-go/tools/clientcmd"
 	"os"
-	"path/filepath"
 	"regexp"
 	"runtime"
 	"testing"
@@ -54,46 +53,6 @@ func TestFlatten(t *testing.T) {
 		assert.NotEmpty(t, auth.ClientKeyData)
 		assert.Equal(t, key, auth.ClientKeyData)
 	}
-
-}
-
-func TestExpandFolder(t *testing.T) {
-
-	currentfolder, err := os.Getwd()
-	parentfolder := filepath.Dir(currentfolder)
-	grandParentFolder := filepath.Dir(parentfolder)
-	baseFolder := "/usr/local/bin"
-	homeFolder, err := os.UserHomeDir()
-	assert.Nil(t, err)
-	expectedHomeFolder := filepath.Join(homeFolder, "tests", "data")
-	expectedParentFolder := filepath.Join(parentfolder, "krknctl")
-	expectedGrandParentFolder := filepath.Join(grandParentFolder, "krknctl")
-	expectedCurrentFolder := filepath.Join(currentfolder, "krknctl")
-	expectedParentFolderWithBase := "/usr/local/krknctl"
-
-	resultParentFolder, err := ExpandFolder("../krknctl", nil)
-	assert.Nil(t, err)
-	assert.Equal(t, expectedParentFolder, *resultParentFolder)
-
-	resultGrandParentFolder, err := ExpandFolder("../../krknctl", nil)
-	assert.Nil(t, err)
-	assert.Equal(t, expectedGrandParentFolder, *resultGrandParentFolder)
-
-	resultCurrentFolder, err := ExpandFolder("./krknctl", nil)
-	assert.Nil(t, err)
-	assert.Equal(t, expectedCurrentFolder, *resultCurrentFolder)
-
-	resultHomeFolder, err := ExpandFolder("~/tests/data", nil)
-	assert.Nil(t, err)
-	assert.Equal(t, expectedHomeFolder, *resultHomeFolder)
-
-	resultSameAbsoluteFolder, err := ExpandFolder(currentfolder, nil)
-	assert.Nil(t, err)
-	assert.Equal(t, currentfolder, *resultSameAbsoluteFolder)
-
-	resultParentFolderWithBase, err := ExpandFolder("../krknctl", &baseFolder)
-	assert.Nil(t, err)
-	assert.Equal(t, expectedParentFolderWithBase, *resultParentFolderWithBase)
 
 }
